# api/docs/openapi.yml
openapi: "3.1.0"
info:
  title: Taillight API
  description: |
    Taillight is a centralized srvlog and application-log viewer with real-time
    streaming, analysis reports, and rsyslog/Juniper integration.
  version: "1.0.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Default API prefix
...
security:
  - session: []
  - apiKey: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /srvlog/{id}:
    get:
      operationId: getSrvlog
      tags: [Srvlog]
      summary: Get single srvlog event
      parameters:
        - $ref: "#/components/parameters/IntID"
      responses:
        "200":
          description: Srvlog event with optional Juniper reference
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/SrvlogEvent"
                      - type: object
                        properties:
                          juniper_ref:
                            type: array
                            items:
                              $ref: "#/components/schemas/JuniperRef"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /stats/summary:
    get:
      operationId: srvlogSummary
      tags: [Stats]
      summary: Srvlog summary
      parameters:
        - $ref: "#/components/parameters/Range"
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/SrvlogSummary"

  # ── Juniper ─────────────────────────────────────────────────────────────
...
  /applog/stats/volume:
    get:
      operationId: applogVolume
      tags: [Stats]
      summary: AppLog volume over time
      parameters:
        - $ref: "#/components/parameters/Interval"
        - $ref: "#/components/parameters/Range"
      responses:
        "200":
          description: Volume buckets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/VolumeBucket"
  /applog/stats/summary:
    get:
      operationId: applogSummary
      tags: [Stats]
      summary: AppLog summary
      parameters:
        - $ref: "#/components/parameters/Range"
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/AppLogSummary"

  # ── Analysis ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tl_session
      description: Session cookie set by POST /auth/login
    apiKey:
      type: http
      scheme: bearer
      description: "API key obtained from POST /auth/keys"

  # ── Parameters ──────────────────────────────────────────────────────────
  parameters:
...
    IntID:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    Interval:
      name: interval
      in: query
      schema:
        type: string
        enum: [1m, 5m, 15m, 30m, 1h, 6h, 1d]
        default: "1h"
    Range:
      name: range
      in: query
      schema:
        type: string
        enum: [1h, 6h, 12h, 24h, 7d, 30d]
        default: "24h"

    # Srvlog-specific
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    SrvlogEvent:
      type: object
      properties:
        id:
          type: integer
        received_at:
          type: string
          format: date-time
        reported_at:
          type: string
          format: date-time
        hostname:
          type: string
        fromhost_ip:
          type: string
        programname:
          type: string
        msgid:
          type: string
        severity:
          type: integer
          minimum: 0
          maximum: 7
        severity_label:
          type: string
          enum: [emerg, alert, crit, err, warning, notice, info, debug]
        facility:
          type: integer
          minimum: 0
          maximum: 23
        facility_label:
          type: string
        syslogtag:
          type: string
        structured_data:
          type: string
          nullable: true
        message:
          type: string
        raw_message:
          type: string
          nullable: true

    # ── AppLog ────────────────────────────────────────────────────────────
...
    VolumeBucket:
      type: object
      properties:
        time:
          type: string
          format: date-time
        total:
          type: integer
        by_host:
          type: object
          additionalProperties:
            type: integer
    SrvlogSummary:
      type: object
      properties:
        total:
          type: integer
        trend:
          type: number
          format: double
          description: Percentage change vs previous period
        errors:
          type: integer
        warnings:
          type: integer
        severity_breakdown:
          type: array
          items:
            type: object
            properties:
              severity:
                type: integer
              label:
                type: string
              count:
                type: integer
              pct:
                type: number
                format: double
        top_hosts:
          type: array
          items:
            $ref: "#/components/schemas/NamedCount"
    AppLogSummary:
      type: object
      properties:
        total:
          type: integer
        trend:
          type: number
          format: double
        errors:
          type: integer
        warnings:
          type: integer
        level_breakdown:
          type: array
          items:
            type: object
            properties:
              level:
                type: string
              count:
                type: integer
              pct:
                type: number
                format: double
        top_services:
          type: array
          items:
            $ref: "#/components/schemas/NamedCount"
    NamedCount:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        pct:
          type: number
          format: double

    # ── Juniper ───────────────────────────────────────────────────────────
    JuniperRef:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        message:
          type: string
        description:
          type: string
        type:
          type: string
        severity:
          type: string
        cause:
          type: string
        action:
          type: string
        os:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Rsyslog ───────────────────────────────────────────────────────────
...